ArticleRepositoryTest 스프링부트 연동
- MemoryArticleRepositoryTest -> ArticleRepositoryTests
- 의존관계 주입을 SpringBoot 연동으로 하여 자동으로 하도록 함
- 테스트를 트랜잭션 단위로 함
